feat(srfi-14): add 17 named char-sets; phase 4 complete
Phase 4 Task 4.2 — final task of Phase 4.

stdlib/lib/srfi/14/named-sets.scm:
- 14 constants via %make-named-charset (letter, lower-case,
  upper-case, title-case, digit, letter+digit, graphic, printing,
  whitespace, iso-control, punctuation, symbol, hex-digit, blank)
- 3 direct constructions (empty, full, ascii)

(srfi 14) library exports: 17 new char-set:* names.

Integration test srfi-14-tests-named-sets.scm — 29 assertions
across 4 groups:
- Basic membership (a/Z/1/space/etc. for each named set)
- Fixed sizes (ascii=128, empty=0, full=#x110000)
- LOAD-BEARING consistency-with-built-in-predicates test:
  walks ASCII range 0..127 and verifies char-set:letter agrees
  with char-alphabetic?, char-set:digit with char-numeric?, etc.
  for all 5 categories (letter/digit/whitespace/lower/upper)
- Non-ASCII Unicode reach (Q4 design): Greek alpha in :letter,
  Cyrillic capital A in :upper-case, Hebrew aleph in :letter
  but not :lower-case or :upper-case

BUG CAUGHT BY CONSISTENCY TEST + FIXED:
extensions/charsets/charsets.go rangeTableToCharSet had uint16
overflow in stride-expansion loop for unicode.Range16. When
r.Lo + r.Stride > 65535, the uint16 loop variable wrapped around
and contaminated the CharSet with spurious low-codepoint values.
Specifically unicode.Lu's [42997, 65313, stride=22316] range
wrapped through 21 (NAK) and 53 ('5'), making them appear in
char-set:upper-case incorrectly. Fix: promote loop variable to
uint32. Affected all 14 FFI-named sets that draw from R16 stride
ranges near the uint16 ceiling. Vindicates the design §10
"consistency with built-in predicates" test.

axis-b-manifest.scm regenerated for the new %make-named-charset
primitive.

Phase 4 exit gate met:
- All 17 named char-sets work
- Consistency with built-in char predicates verified across ASCII
- Non-ASCII Unicode reach (Q4) verified for Greek/Cyrillic/Hebrew
- 91 cumulative SRFI-14 integration assertions pass
- make lint && make covercheck both clean

+++ extensions/charsets/charsets.go
@@ -810,6 +810,10 @@ func makeNamedCharSet(name string) (*values.CharSet, error) {
 // Handles both Range16 (BMP, codepoints ≤ 0xFFFF) and Range32 (supplementary
 // planes). Stride > 1 means the range covers every Stride-th codepoint;
 // expand into per-stride unit ranges.
+//
+// NOTE: the loop variable must be promoted to uint32 for R16 ranges. Using
+// uint16 directly (the element type of Range16 fields) overflows when
+// hi + stride > 65535, producing wraparound values that contaminate the set.
 func rangeTableToCharSet(t *unicode.RangeTable) *values.CharSet {
 	var rs []values.CharSetRange
 	for _, r := range t.R16 {
@@ -817,7 +821,8 @@ func rangeTableToCharSet(t *unicode.RangeTable) *values.CharSet {
 			rs = append(rs, values.CharSetRange{Lo: rune(r.Lo), Hi: rune(r.Hi)})
 			continue
 		}
-		for cp := r.Lo; cp <= r.Hi; cp += r.Stride {
+		// Promote to uint32 to avoid uint16 wraparound when lo + stride > 65535.
+		for cp := uint32(r.Lo); cp <= uint32(r.Hi); cp += uint32(r.Stride) {
 			rs = append(rs, values.CharSetRange{Lo: rune(cp), Hi: rune(cp)})
 		}
 	}
